feat: refactor item + map + session flow, add session_manager, remove old validators
- Reworked item pipeline (~40% cleaner logic)
- Updated map & zone services for new session flow
- Added session_manager for handling old/new session kick
- Removed deprecated item_validator & message_service
- Synced player/mob services with new data structures

+++ src/entities/item_template.rs
@@ -6,7 +6,7 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "item_template")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub id: i32,
+    pub id: i16,
     #[sea_orm(column_name = "TYPE")]
     pub r#type: i32,
     pub gender: i16,
